Add support for even more lemmy v4 endpoints

+++ api/src/main/kotlin/com/idunnololz/summit/api/dto/lemmy/v4/models/HideCommunity.kt
@@ -37,7 +37,7 @@ data class HideCommunity (
 
     /* The community id. */
     @SerializedName("community_id")
-    val communityId: kotlin.Double
+    val communityId: kotlin.Int
 
 ) {
