+++ api/src/main/kotlin/com/idunnololz/summit/api/dto/lemmy/v4/models/PurgeComment.kt
@@ -33,7 +33,7 @@ data class PurgeComment (
 
     /* The comment id. */
     @SerializedName("comment_id")
-    val commentId: kotlin.Double
+    val commentId: kotlin.Int
 
 ) {
