+++ api/src/main/kotlin/com/idunnololz/summit/api/dto/lemmy/v4/models/PurgeCommunity.kt
@@ -33,7 +33,7 @@ data class PurgeCommunity (
 
     /* The community id. */
     @SerializedName("community_id")
-    val communityId: kotlin.Double
+    val communityId: kotlin.Int
 
 ) {
