+++ api/src/main/kotlin/com/idunnololz/summit/api/dto/lemmy/v4/models/PurgePerson.kt
@@ -33,7 +33,7 @@ data class PurgePerson (
 
     /* The person id. */
     @SerializedName("person_id")
-    val personId: kotlin.Double
+    val personId: kotlin.Long
 
 ) {
